refactor: ui/component のパッケージ構成を修正
